requirements.txt
altgraph==0.17.4
macholib==1.16.3
packaging==24.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-dotenv==1.0.1
setuptools==75.1.0
